test(browser): wait for license check to complete
git:da5de01341cb4cd7938d9ebb33fec81dd4574d15 introduced a regression in
the password tests. Previously the waiting for the license check was
necessary since otherwise sometimes the password was changed before this
check completed, leading to an error.

Add it back here only for the password test, since it actually did
improve the stability of the none password tests.

Bug #56054