univention-samba4-backup: don't touch backups every day
The backups get their permissions changed every time the script is running. This changes the ctime of the file and causes inclusion to systembackups (e.g. bareos). This is especially annoying when many backupfiles are kept or the files are large.

In f065824d7a9b091dfa8c425da5006123770f760c calling "chmod" for every files was added, now we do it only for files with wrong permissions.