fix(join): remove shell glob from rsync command
The '*' at the end of the rsync command to copy SSL certificates is
unnecessary, since a trailing / in rsync does the same thing; copy all
the content of the directory.

This caused issues with the /etc/univention/ssl having too many files
and the list of arguments to rsync becoming too long.

Bug #57421