Revert "fix(univention-keycloak): check if mapper exists that maps uid"
This reverts (parts of) commit 6827db2046158906888f7a2317d8bc9373633e44.

Only revert the ucs-test part. It turned out to be relatively flaky and
the implications of just installing keycloak are not fully known and
might lead to errors.

Initially this commit stemmed out of not knowing, that we do have tests,
testing the SAML login on the UMC for Keycloak, so no test coverage is
actually being removed.

Bug #57420