Bug #57482: add an SQL database for shared session storage
The OIDC backchannel logout can hit any of the UMCs if multiprocessing
is configured. This commit introduces a persistent storage for sessions
between UMC processes.

Currently only the sessionId and various relevant OIDC token claims are
saved. This makes it not possible to use sessions between multiple UMC
processes but rather just adds the ability for other UMC processes to
delete the sessions of others.

Note: currently, if backchannel logout request is handled by a process
that is not the one it originated from the automatic tab refresh on
logout does not work.

Also add a test that tests OIDC backchannel logout with multiprocessing
enabled.

univention/ucs#2371