Issue #245: improve Guardian branch tests
univention/components/authorization-engine/guardian#245

There are two key features that are missing from the branch tests, which
prevent a full testing of the branch before merge:

* The ability to set an Appcenter version that matches with the corresponding
  changes to the Docker image.
* The Guardian Management UI also should be installed.

This commit updates the guardian branch test scenario and supporting scripts
to make this happen.

One change that was necessary was to replace `exec` in the script with `eval`.
If the `univention-app install` script has any issues, it exits the whole
shell, even if the install was successful. In this case:

```
Executing interface update_available for guardian-authorization-api
No interface defined
```

This error is unrelated to the current MR and does not affect the installation
of the guardian.

The `eval` command prevents the script from exiting before it is complete.