fix(umc, Bug #57533): update global bind function in UMC UDM module
there is a global bind function that will be used for certain requests
in the UMC UDM module. This bind function was previously only set when
the module initialized, which was fine since a user password remained
the same and whenever a SAML assertion expired, a new session would be
created, and with that new UMC modules.

However with OIDC the access token changes frequently (by default ever 5
minutes).
When the access token expires and therefore gets refreshed the UDM
module will use the old access token since it's set only on initial
module creation.

This commit adds a way to update this global bind function by checking,
before each request, if the bindhash changed. If it did, we use the
requests bind function as the new global bind function. This requests
bind function will contain an updated access token.