fix (umc, Bug #57516): make OIDC front-channel logout work
In environments where the UMC uses OIDC with an OpenID Provider from a
separate domain, the HTTP Content-Security-Policy frame-ancestors directive
must be used to allow the OpenID Provider to embed the hidden iframe which is
used during front-channel logout to navigate back to the RP.

The changes to oidc.py are made because there was a flaw in the
implementation of the front-channel logout flow. Previously the UMC's
`oidc/frontchannel-logout` endpoint made an unnecessary redirect that is now
removed. See univention/ucs#2361 / Bug #57466 for more information.

Issue univention/ucs#2399