fix: add a LRU cache for the most recent DNs (Issue univention/ucs#2653)
The DN class is merely only for unifying and checking DN strings and
the conversion between DN objects and DN strings. The same input results
in the same output. Since the conversion is expensive, we should
cache results.
This is especially helpful in case UDM works on a similar set of DNs
(e.g. uniqueMember entries when modifying groups which are mostly the
same in "old" and "new")

TODO FIXME the max size of the LRU cache has to be defined.
It would be best to start with a good default for most envs and give
the opportunity to change the value via UCR.