fix: useless use of keepCase() (Issue univention/ucs#2653)
In the beginning of _ldap_modlist() the DNs of users/hosts/nestedGroups
get unified by DN.set(...). Afterwards uniqueMemberRemove is evaluated
via "old-new". keepCase tries to reuse the same case as in the original
in old. Therefore this call is useless and can be dropped.

TODO: we have to check if this is correct or if the evaluation of
uniqueMemberRemove is already a mistake.