Bug #58183: fix tests for Keycloak 26
The browser gets started in German due to tests being started with
LANG=de_DE.UTF-8 in utils.sh, however UCSBrowserTest ignores this
variable. An attempt was made to sync the LANG env and the
UCSBrowserTest language however this revealed certain issues with the
translation.

For now just run all the keycloak browser tests in English