ci(gitlab): remove packages from pipelines where no changes exist
pipelines for merge requests, where changes to a package were done in a
recent commit but in the latest commit the changes have been reverted,
so that no changes exists when comparing with the target branch, never
removed the stale packages from the apt repository

A new job was added which checks which unchanges packages needs to be removed upstream.

unfixed: there was a race condition, if you push multiple times to a repo:
the build which is faster wins. even if it is the older pipeline.
this happens e.g. if you have mixed changes for packages which are fast and slow
in build-time and you remove the whole changes for a slow build / introduce a
new slow build, the older wrong pipeline might win.

additional new race condition:
if you push multiple times and the "drop unused packages" is slower than the new
pipeline, which already created packages for newly-changed packages, they will be removed.
Handling this would require to request the gitlab API and cancel all prior running jobs.
The drop job is at least executed prior building new packages.

Bug #42294