feat(ad-connector): Allow synchronization of subtree ancestors
When working with more complicated LDAP structures, it is possible
that a specific subtree should be synchronized which is a few levels
below the directory root. Currently, this can be accomplished by
allowing an ancestor tree and working with the UCR-V ignoresubtree.
This can be tedious to setup, as many entries in the ignoresubtree UCRV
are needed and maintenance is required as new neighbors get created.
This commit introduces the UCR-V con.*/ad/mapping/allow-subtree-ancestors
which, if enabled, will allow objects which are ancestors to the
root of the subtrees set in connector/ad/mapping/allowsubtree/.*/[ad|ucs]
to be synchronized.

Bug #57979
Issue univention/dev/ucs#2834