fix(umc): fix auth/faillock for UMC pam stack
`pam_faillock.so authfail` is used to actually record a failed
authentication attempt. It was missing in the
`univention-management-console` PAM stack and has now been added.

Furthermore change `pam_faillock.so preauth` from `required` to
`requisite` to better align with how the stack terminates when the
global account locking is turned on.

Bug: 57968