test(keycloak): add test for case sensitive username in subject OIDC field
A recent Keycloak upgrade changed how the representation of the
username is send in the `sub` field of the OIDC. This caused a
pullcord.

This test ensures that the username is contained in the token’s
OIDC subject (sub) claim, and it preserves case sensitivity.

univention/dev/internal/dev-issues/dev-incidents#186

Bug #58805